% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{The CountryPopulation Function}
\usage{
CountryPopulation(CountryInput)
}
\arguments{
\item{CountryInput}{Name of the country to plot}
}
\value{
A graph of the country's population from 1950 to 2020.
}
\description{
A function that when given a Country's name produces
a graph of the population from 1950 to 2020
}
\examples{
{
CountryPopulation("China")
CountryPopulation("Brazil")
}
}
