% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Cup.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{A data set on World Cup Attendance.}
\format{
A data frame with 22 observations of 4 variables
\describe{
  \item{Totalattendance}{The total attendance for the World Cup for a given year. }
  \item{Matches}{The number of matches played in the World Cup for a given year.}
  \item{Averageattendance}{The average attendance for the World Cup for a given year.}
  \item{WorldCup}{The name of the country that hosted the World Cup and the year of the World Cup.}
}
}
\source{
\url{"https://en.wikipedia.org/wiki/FIFA_World_Cup"}
}
\usage{
World_Cup
}
\description{
A data set on World Cup Attendance.
}
\keyword{datasets}
